[fix] resolve GC Bug
Issue 1: When adding a new object to the constant table, if the table requires more space, the garbage collector may mistakenly sweep the new object before it's added.

Issue 2: When adding a new string to the string table, if additional space is needed, the garbage collector may sweep the new string prematurely.